+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Code style
 - Update Angular to 7 and packager
 - Make timer more accurate
+- Extend API by `setCustomActivityEvents(customEvents: Observable<any>)`
 
 ## 2.0.0
 - Replace rollup by Angular CLI 6 bundling.